requirements.txt
APScheduler==3.11.0
Bootstrap_Flask==2.2.0
Flask==2.3.2
Flask_Login==0.6.3
flask_sqlalchemy==3.1.1
itsdangerous==2.2.0
python-dotenv==1.1.1
Requests==2.32.4
SQLAlchemy==2.0.25
Werkzeug==3.1.3
gunicorn==23.0.0
